Rename eh1 file and mirror for eh2
