+++ src/i19_bluesky/eh1/__init__.py
@@ -1,6 +1,6 @@
 """A place for eh1 specific plans."""
 
-from .optics_hutch_control import close_experiment_shutter, open_experiment_shutter
+from .eh1_optics_hutch_control import close_experiment_shutter, open_experiment_shutter
 
 __all__ = [
     "open_experiment_shutter",
