+++ src/i19_bluesky/eh2/__init__.py
@@ -1,1 +1,5 @@
 """A place for eh2 specific plans."""
+
+from .eh2_optics_hutch_control import close_experiment_shutter, open_experiment_shutter
+
+__all__ = ["open_experiment_shutter", "close_experiment_shutter"]
